Add support for disabling backend in init, clean up tests.
This change adds support for the `-backend=false` init argument,
used to disable a configured backend. It also moves init inside
the `setup()` method so it's run regardless of the value passed in
the `command` argument. Minimal tests cleanup is also in this
change, removing `nose`-specific imports and making it possible to
use `pytest` as a test runner.