Make cache directory path Windows friendly (#71)
* Make cache directory path Windows friendly

The existing logic for building the cache directory path assumes *nix style paths when attempting to make a unique cache dir for each test fixture and step. Paths with a Windows drive prefix on the front ("C:\...") muck this up. Pathlib's '/' operator silently covers this up (under Python 3.9 on Windows, at least).

This change preserves the uniqueness logic whilst being Window friendly by taking a hash of the path to the test fixture's configuration directory `tfdir` and using that in the cache directory path construction.

* Fix linting errors

* Fix linting errors